rename the source dir
